Moved posts filter and fetch logic to separate custom hook: usePostsFilter.
